start hash work
