+++ loc_chain_app/lib/util/transaction.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/services.dart';
+import 'package:fast_rsa/model/bridge.pb.dart';
+import 'package:fast_rsa/rsa.dart';
+
+class Transaction {}
